Kelp_ai_teaser 2/requirements.txt
google-generativeai
python-pptx
python-docx
beautifulsoup4
requests
presidio-analyzer
presidio-anonymizer
spacy
# Run this after installing requirements:
# python -m spacy download en_core_web_lg
